+++ CHANGELOG.md
@@ -227,6 +227,7 @@ Version 1.3.1 - New features, Refactors, Main improvements and User friendly imp
 - [x] Add lock button [#88](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/88) into equations manage panel
 - [x] Add show headings only panel, so in outline view, we can just use it as outline panel.
 - [x] Optimize `renderImageCaptions` function, now it will not parse Markdown every time when fast typing.
+- [x] Add enhancement [#91](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/91), stop auto-numbering when theres content after equation block.